fix: shrink the span unit kwargs

+++ crates/ryo3-jiff/src/ry_offset.rs
@@ -297,7 +297,7 @@ impl RyOffset {
     fn add(
         &self,
         other: Option<Spanish>,
-        hours: i64,
+        hours: i32,
         minutes: i64,
         seconds: i64,
         milliseconds: i64,
@@ -344,7 +344,7 @@ impl RyOffset {
     fn sub(
         &self,
         other: Option<Spanish>,
-        hours: i64,
+        hours: i32,
         minutes: i64,
         seconds: i64,
         milliseconds: i64,
