+++ crates/ryo3-jiff/src/ry_time.rs
@@ -177,7 +177,7 @@ impl RyTime {
     fn add(
         &self,
         other: Option<Spanish>,
-        hours: i64,
+        hours: i32,
         minutes: i64,
         seconds: i64,
         milliseconds: i64,
@@ -237,7 +237,7 @@ impl RyTime {
     fn sub(
         &self,
         other: Option<TemporalSubInput<Self>>,
-        hours: i64,
+        hours: i32,
         minutes: i64,
         seconds: i64,
         milliseconds: i64,
